requirements.txt
alabaster==0.7.12
apipkg==1.5
appdirs==1.4.3
application-utility==1.1.0
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
btrfsutil==1.1.0
CacheControl==0.12.5
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
colorama==0.4.1
cryptography==2.6.1
cupshelpers==1.0
distlib==0.2.8
distro==1.4.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==0.10.1
freezegun==0.3.11
future==0.17.1
gufw==18.10.0
html5lib==1.0.1
idna==2.7
imagesize==1.1.0
iniconfig==1.0.0
isort==4.3.17
itsdangerous==0.24
Jade-Application-Kit==1.3.5
Jinja2==2.8
jsonschema==2.6.0
lazy-object-proxy==1.3.1
lockfile==0.12.2
louis==3.9.0
MarkupSafe==0.23
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
msgpack==0.6.1
npyscreen==4.10.5
packaging==19.0
pacman-mirrors==4.14.2
paramiko==2.4.2
pbr==5.1.3
pep517==0.5.0
Pillow==6.0.0
pluggy==0.9.0
pretend==1.0.9
progress==1.5
pwquality==1.4.0
py==1.8.0
pyasn1==0.4.5
pycairo==1.18.0
pycparser==2.19
pycups==1.9.74
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.32.1
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.4.0
PyQt5==5.12.1
PyQt5-sip==4.19.16
pytest==4.4.1
pytest-runner==4.4
python-dateutil==2.8.0
python-telegram-bot==12.0.0b1
pytoml==0.1.20
pytz==2019.1
pyxdg==0.26
PyYAML==3.13
requests==2.20.1
retrying==1.3.3
scripttest==1.3
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
team==1.0
texttable==0.9.1
tornado==6.0.2
typed-ast==1.3.5
ufw==0.36
urllib3==1.24.2
virtualenv==16.1.0
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.10.4
wrapt==1.11.1
